指摘点の改善
