+++ app/controllers/users_controller.rb
@@ -1,3 +1,4 @@
+
 class UsersController < ApplicationController
   def new
     @user = User.new
